fix: add workaround for race in SelectorEventLoop sock_connect
Workaround the race documented in https://github.com/python/cpython/issues/131728
