+++ tests/test_impl.py
@@ -14,48 +14,7 @@ from aiohappyeyeballs import (
     start_connection,
 )
 
-
-def mock_socket_module():
-    m_socket = mock.MagicMock(spec=socket)
-    for name in (
-        "AF_INET",
-        "AF_INET6",
-        "AF_UNSPEC",
-        "IPPROTO_TCP",
-        "IPPROTO_UDP",
-        "SOCK_STREAM",
-        "SOCK_DGRAM",
-        "SOL_SOCKET",
-        "SO_REUSEADDR",
-        "inet_pton",
-    ):
-        if hasattr(socket, name):
-            setattr(m_socket, name, getattr(socket, name))
-        else:
-            delattr(m_socket, name)
-
-    m_socket.socket = mock.MagicMock()
-    m_socket.socket.return_value = mock_nonblocking_socket()
-
-    return m_socket
-
-
-def mock_nonblocking_socket(
-    proto=socket.IPPROTO_TCP, type=socket.SOCK_STREAM, family=socket.AF_INET
-):
-    """Create a mock of a non-blocking socket."""
-    sock = mock.create_autospec(socket.socket, spec_set=True, instance=True)
-    sock.proto = proto
-    sock.type = type
-    sock.family = family
-    sock.gettimeout.return_value = 0.0
-    return sock
-
-
-def patch_socket(f):
-    return mock.patch("aiohappyeyeballs.impl.socket", new_callable=mock_socket_module)(
-        f
-    )
+from .conftest import patch_socket
 
 
 @pytest.mark.asyncio
@@ -1520,289 +1479,6 @@ async def test_all_same_exception_and_with_different_errno(
 
 @patch_socket
 @pytest.mark.asyncio
-async def test_uvloop_runtime_error(
-    m_socket: ModuleType,
-) -> None:
-    """
-    Test RuntimeError is handled when connecting a socket with uvloop.
-
-    Connecting a socket can raise a RuntimeError, OSError or ValueError.
-
-    - OSError: If the address is invalid or the connection fails.
-    - ValueError: if a non-sock it passed (this should never happen).
-    https://github.com/python/cpython/blob/e44eebfc1eccdaaebc219accbfc705c9a9de068d/Lib/asyncio/selector_events.py#L271
-    - RuntimeError: If the file descriptor is already in use by a transport.
-
-    We should never get ValueError since we are using the correct types.
-
-    selector_events.py never seems to raise a RuntimeError, but it is possible
-    with uvloop. This test is to ensure that we handle it correctly.
-    """
-    mock_socket = mock.MagicMock(
-        family=socket.AF_INET,
-        type=socket.SOCK_STREAM,
-        proto=socket.IPPROTO_TCP,
-        fileno=mock.MagicMock(return_value=1),
-    )
-    create_calls = []
-
-    def _socket(*args, **kw):
-        for attr in kw:
-            setattr(mock_socket, attr, kw[attr])
-        return mock_socket
-
-    async def _sock_connect(
-        sock: socket.socket, address: Tuple[str, int, int, int]
-    ) -> None:
-        create_calls.append(address)
-        raise RuntimeError("all fail")
-
-    m_socket.socket = _socket  # type: ignore
-    ipv6_addr_info = (
-        socket.AF_INET6,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("dead:beef::", 80, 0, 0),
-    )
-    ipv6_addr_info_2 = (
-        socket.AF_INET6,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("dead:aaaa::", 80, 0, 0),
-    )
-    ipv4_addr_info = (
-        socket.AF_INET,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("107.6.106.83", 80),
-    )
-    addr_info = [ipv6_addr_info, ipv6_addr_info_2, ipv4_addr_info]
-    local_addr_infos = [
-        (
-            socket.AF_INET6,
-            socket.SOCK_STREAM,
-            socket.IPPROTO_TCP,
-            "",
-            ("::1", 0, 0, 0),
-        ),
-        (
-            socket.AF_INET,
-            socket.SOCK_STREAM,
-            socket.IPPROTO_TCP,
-            "",
-            ("127.0.0.1", 0),
-        ),
-    ]
-    loop = asyncio.get_running_loop()
-    # We should get the same exception raised if they are all the same
-    with mock.patch.object(loop, "sock_connect", _sock_connect), pytest.raises(
-        RuntimeError, match="all fail"
-    ):
-        assert (
-            await start_connection(
-                addr_info,
-                happy_eyeballs_delay=0.3,
-                interleave=2,
-                local_addr_infos=local_addr_infos,
-            )
-            == mock_socket
-        )
-
-    # All calls failed
-    assert create_calls == [
-        ("dead:beef::", 80, 0, 0),
-        ("dead:aaaa::", 80, 0, 0),
-        ("107.6.106.83", 80),
-    ]
-
-
-@patch_socket
-@pytest.mark.asyncio
-async def test_uvloop_different_runtime_error(
-    m_socket: ModuleType,
-) -> None:
-    """Test different RuntimeErrors are handled when connecting a socket with uvloop."""
-    mock_socket = mock.MagicMock(
-        family=socket.AF_INET,
-        type=socket.SOCK_STREAM,
-        proto=socket.IPPROTO_TCP,
-        fileno=mock.MagicMock(return_value=1),
-    )
-    create_calls = []
-    counter = 0
-
-    def _socket(*args, **kw):
-        for attr in kw:
-            setattr(mock_socket, attr, kw[attr])
-        return mock_socket
-
-    async def _sock_connect(
-        sock: socket.socket, address: Tuple[str, int, int, int]
-    ) -> None:
-        create_calls.append(address)
-        nonlocal counter
-        counter += 1
-        raise RuntimeError(counter)
-
-    m_socket.socket = _socket  # type: ignore
-    ipv6_addr_info = (
-        socket.AF_INET6,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("dead:beef::", 80, 0, 0),
-    )
-    ipv6_addr_info_2 = (
-        socket.AF_INET6,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("dead:aaaa::", 80, 0, 0),
-    )
-    ipv4_addr_info = (
-        socket.AF_INET,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("107.6.106.83", 80),
-    )
-    addr_info = [ipv6_addr_info, ipv6_addr_info_2, ipv4_addr_info]
-    local_addr_infos = [
-        (
-            socket.AF_INET6,
-            socket.SOCK_STREAM,
-            socket.IPPROTO_TCP,
-            "",
-            ("::1", 0, 0, 0),
-        ),
-        (
-            socket.AF_INET,
-            socket.SOCK_STREAM,
-            socket.IPPROTO_TCP,
-            "",
-            ("127.0.0.1", 0),
-        ),
-    ]
-    loop = asyncio.get_running_loop()
-    # We should get the same exception raised if they are all the same
-    with mock.patch.object(loop, "sock_connect", _sock_connect), pytest.raises(
-        RuntimeError, match="Multiple exceptions: 1, 2, 3"
-    ):
-        assert (
-            await start_connection(
-                addr_info,
-                happy_eyeballs_delay=0.3,
-                interleave=2,
-                local_addr_infos=local_addr_infos,
-            )
-            == mock_socket
-        )
-
-    # All calls failed
-    assert create_calls == [
-        ("dead:beef::", 80, 0, 0),
-        ("dead:aaaa::", 80, 0, 0),
-        ("107.6.106.83", 80),
-    ]
-
-
-@patch_socket
-@pytest.mark.asyncio
-async def test_uvloop_mixing_os_and_runtime_error(
-    m_socket: ModuleType,
-) -> None:
-    """Test uvloop raising OSError and RuntimeError."""
-    mock_socket = mock.MagicMock(
-        family=socket.AF_INET,
-        type=socket.SOCK_STREAM,
-        proto=socket.IPPROTO_TCP,
-        fileno=mock.MagicMock(return_value=1),
-    )
-    create_calls = []
-    counter = 0
-
-    def _socket(*args, **kw):
-        for attr in kw:
-            setattr(mock_socket, attr, kw[attr])
-        return mock_socket
-
-    async def _sock_connect(
-        sock: socket.socket, address: Tuple[str, int, int, int]
-    ) -> None:
-        create_calls.append(address)
-        nonlocal counter
-        counter += 1
-        if counter == 1:
-            raise RuntimeError(counter)
-        raise OSError(counter, f"all fail {counter}")
-
-    m_socket.socket = _socket  # type: ignore
-    ipv6_addr_info = (
-        socket.AF_INET6,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("dead:beef::", 80, 0, 0),
-    )
-    ipv6_addr_info_2 = (
-        socket.AF_INET6,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("dead:aaaa::", 80, 0, 0),
-    )
-    ipv4_addr_info = (
-        socket.AF_INET,
-        socket.SOCK_STREAM,
-        socket.IPPROTO_TCP,
-        "",
-        ("107.6.106.83", 80),
-    )
-    addr_info = [ipv6_addr_info, ipv6_addr_info_2, ipv4_addr_info]
-    local_addr_infos = [
-        (
-            socket.AF_INET6,
-            socket.SOCK_STREAM,
-            socket.IPPROTO_TCP,
-            "",
-            ("::1", 0, 0, 0),
-        ),
-        (
-            socket.AF_INET,
-            socket.SOCK_STREAM,
-            socket.IPPROTO_TCP,
-            "",
-            ("127.0.0.1", 0),
-        ),
-    ]
-    loop = asyncio.get_running_loop()
-    # We should get the same exception raised if they are all the same
-    with mock.patch.object(loop, "sock_connect", _sock_connect), pytest.raises(
-        OSError, match="Multiple exceptions: 1"
-    ):
-        assert (
-            await start_connection(
-                addr_info,
-                happy_eyeballs_delay=0.3,
-                interleave=2,
-                local_addr_infos=local_addr_infos,
-            )
-            == mock_socket
-        )
-
-    # All calls failed
-    assert create_calls == [
-        ("dead:beef::", 80, 0, 0),
-        ("dead:aaaa::", 80, 0, 0),
-        ("107.6.106.83", 80),
-    ]
-
-
-@patch_socket
-@pytest.mark.asyncio
 async def test_handling_system_exit(
     m_socket: ModuleType,
 ) -> None:
@@ -1978,7 +1654,6 @@ async def test_cancellation_is_not_swallowed(
     "connect_side_effect",
     [
         OSError("during connect"),
-        asyncio.CancelledError("during connect"),
     ],
 )
 @patch_socket
@@ -2014,3 +1689,4 @@ async def test_single_addr_info_close_errors(
     ]
     with pytest.raises(OSError, match="during close"):
         await start_connection(addr_info)
+    await asyncio.sleep(0)
